Add break penalties and money precision
Adds configurable foreign-claim break penalties, including bank charges, damage handling, and uniform mining timing. Also upgrades bank balances and config values to support decimal money, adds NBT coercion for legacy saves, and wires client/server mining penalty packets for synced break timing.

+++ factions-1.20.1_BLIS/src/main/java/io/icker/factions/config/BankConfig.java
@@ -13,5 +13,5 @@ public class BankConfig {
     public String _COMMENT_MAX_BALANCE = "Maximum bank balance per faction (-1 = unlimited)";
 
     @SerializedName("maxBalance")
-    public int MAX_BALANCE = -1;
+    public double MAX_BALANCE = -1;
 }
